Tast 3: droprate Issue/ Task 2 recommit
For task 2 the collider was not correctly assigned in the editor so i had to recommit.
Task 3 issue: droprates were wrong and the dropchance was not 100%.
The SelectDropRandomly method in the LootDesctiption class was the issue. It had an for-loop iterating over the various loot types. But once it was finished, if no drop was successfully found it would return null; I changed the for loop to be inside a while loop so if during the first for run no drop was selected, the loop would be run again until a drop was found.  So it is not possible to have no loot anymore.