Task 5: Iterate 2 completed
Task: add an outline to the ability when charging and scale it up, scale it back down when on cooldown.
Added an outline module to the icon that gets enable when the ability is getting charged and gets disabled when finished charging. a coroutine get sused to scale the ability UI