Task 6: completed
Task:  create a second ability that can be dropped by lizards.
I reformated parts of the code to be reusable for a second skill. the drops have a new enum which can be used to define if the drop should be a collectable or an ability, if it is an ability a scriptable object with the description of the ability (similar to the BasicAttack) can be added. this will be the new ability that the player can pick up.
To create new abilities, we only need to create a new drop scriptable object as well as a new prefab to link the new spell to.